Utility class for loading resources into byte array.
Created a Utility singleton class with a function that will extract the byte array from a specified file path in a specified directory. Also created unit tests for the function for when a resource doesn't exist, when a nil file path is passed in, and when the resource does exist.

Initial Utility class for generic utility functions. This may need to be moved to a generic Resource Utility class.

Resolves: #6662