Implement Player class and basic Hub view
The Hub implementation is split into a view and a view controller.
The view handles the front-end elements, while the view controller handles the logic.
The Hub only has a gold generation button that, when pressed, increments the player's gold by 5 and places the button on a 5 second cooldown timer.

The Player class doesn't support holding items yet.
For now, the Player instance is held in The Hub's view controller. This may be moved once data needs to be shared between views.