Basic GL Renderer
Created a Renderer class that compiles a vertex and fragment shader compiles and links them into a program. Can get and set uniforms for the program. Created a basic vertex and fragment shader, created a Mesh class that should be able to hold the data we need for a Mesh.

TODO:
It doesn't break but it doesn't show anything, also no errors.
Comment the Render code.