Implement high scores handling
Scores are based on clear time per dungeon. Time is calculated only during combat, and not when a node is cleared.
Shorter times are better scores. Only the "best" time is stored.
"Best times" are stored per dungeon and are displayed in the Dungeons view, when a dungeon is selected.

Resolves: #6901